chore: move errno to shared
