--- server/cmd/api/biz/errno/Makefile
@@ -1,2 +1,0 @@
-server:
-	kitex -module github.com/CyanAsterisk/FreeCar  ./../../../../idl/errno.thrift
